requirements.txt
asgiref==3.8.1
Django==5.0.6
django-htmx==1.17.3
gunicorn==22.0.0
packaging==24.0
psycopg==3.1.19
psycopg-binary==3.1.19
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.0
